Trello Stories Add haptic feedback Find ATM page by default shows a list of “Found 0 results” -- This should be hidden when I first go to the atm page and only show up once I actually search for something If the card is pending, we are unable to lock or unlock or change preferences —> we should have a visual indicator of it being in pending status So that the user knows why he/she is unable to tap -- design update required Popup for first time user experience

+++ NorthLoopFin/NorthLoopFin/Application Views/Home/Models/AccountsInfo/AccountBalance.swift
@@ -9,7 +9,7 @@
 import Foundation
 
 struct AccountBalance: Codable {
-    let amount: Int
+    let amount: Double
     let currency: String
     
     enum CodingKeys: String, CodingKey {
@@ -19,7 +19,7 @@ struct AccountBalance: Codable {
     
     init(from decoder: Decoder) throws {
         let values = try decoder.container(keyedBy: CodingKeys.self)
-        amount = try values.decodeIfPresent(Int.self, forKey: .amount) ?? 0
+        amount = try values.decodeIfPresent(Double.self, forKey: .amount) ?? 0.0
         currency = try values.decodeIfPresent(String.self, forKey: .currency) ?? ""
     }
 }
